refactor:파일구조에 따른 파일 이동
